[maint] Move code in "activity" to src dir
