+++ lib/src/trial_response/trial_response_controller.dart
