+++ lib/src/trial_stim/stim_controller.dart
